Adjust SECCOMP BPF policy
Allow the following syscalls:

- fcntl
- fstat
- lseek
- ioctl

All these syscalls are allowed in the FreeBSD capability mode sandbox,
however we are also in a position to do additional filtering based on
syscall arguments.  In most cases, these syscalls take in bitmasks
and integer flags, therefore we should be able to apply additional
filtering.